% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald_RcppParallel.R
\name{get_cps}
\alias{get_cps}
\title{get change point estimates}
\usage{
get_cps(stat, thresh, G, nu = 1/4, criterion = c("eps", "eta"))
}
\description{
get change point estimates
}
\keyword{internal}
