% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.R
\name{MBS_RECUR}
\alias{MBS_RECUR}
\title{binary segmentation inner loop}
\usage{
MBS_RECUR(
  x,
  order,
  p,
  s,
  e,
  thresh,
  G,
  estim = "C",
  var_estim = "Local",
  cps,
  stat = list(),
  iter = 1,
  criterion = "eps",
  nu = 0.25,
  alpha = 0.05
)
}
\description{
binary segmentation inner loop
}
\keyword{internal}
