% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiscale.R
\name{MFA}
\alias{MFA}
\title{MOSUM procedure for multiple time series at multiple scales}
\usage{
MFA(
  x,
  p,
  Gset,
  test = c("Wald", "Score")[1],
  estim = c("DiagC", "DiagH")[1],
  alpha = 0.05
)
}
\arguments{
\item{x}{data matrix}

\item{p}{integer VAR model order}

\item{Gset}{integer vector of MOSUM bandwidths}

\item{estim}{string estimation method}

\item{alpha}{Numeric significance level}

\item{method}{string indicating which of `Wald` or `Score` to use}

\item{varEstim}{string variance estimation method}
}
\value{
list containing Boolean test outcome `Reject`, Numeric rejection threshold `Threshold`, 
 Numeric vector of test statistic `mosum`, Integer vector of estimated changepoints `cps`, Plot `plot`, 
 String of input estimator `estim`
}
\description{
MOSUM procedure for multiple time series at multiple scales
}
\examples{
data(voldata)
MFA(voldata[,2:5], 1, c(100, 250, 400) )
}
