% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum_lm.R
\name{mosum_lm_bs}
\alias{mosum_lm_bs}
\title{MOSUM Binary Segmentation procedure for multivariate regression}
\usage{
mosum_lm_bs(X, G, alpha = 0.05, max_iter = 10)
}
\arguments{
\item{X}{data matrix with response in column 1, and intercept in any other column}

\item{G}{integer MOSUM bandwidth}

\item{alpha}{Numeric significance level}

\item{max_iter}{Integer maximum number of Binary Segmentation recursions to allow}
}
\value{
list containing Boolean test outcome `Reject`, Numeric rejection threshold `Threshold`, 
 Numeric vector of test statistic `mosum`, Integer vector of estimated change points `cps`, Plot `plot`,
}
\description{
MOSUM Binary Segmentation procedure for multivariate regression
}
\examples{
data(X0df)
mosum_lm_bs(X0df, 200)
data(X1df)
mosum_lm_bs(X1df, 200)
}
