% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{mosum_sub}
\alias{mosum_sub}
\title{MOSUM subsampling procedure for multiple time series}
\usage{
mosum_sub(
  x,
  p,
  G,
  method = c("Wald", "Score")[1],
  estim = c("DiagC", "DiagH")[1],
  varEstim = c("Local", "Global")[1],
  kap = 1,
  alpha = 0.05,
  criterion = c("eps", "eta")[1],
  nu = 0.25
)
}
\arguments{
\item{x}{data matrix}

\item{p}{integer VAR model order}

\item{G}{integer MOSUM bandwidth}

\item{method}{string indicating which of `Wald` or `Score` to use}

\item{estim}{string estimation method}

\item{varEstim}{string variance estimation method}

\item{kap}{Numeric subsampling resolution constant}

\item{alpha}{Numeric significance level}

\item{criterion}{string location procedure}

\item{nu}{Numeric location procedure hyperparameter}
}
\value{
list containing Boolean test outcome `Reject`, Numeric rejection threshold `Threshold`, 
 Numeric vector of test statistic `mosum`, Integer vector of estimated changepoints `cps`, Plot `plot`, 
 String of input estimator `estim`
}
\description{
MOSUM subsampling procedure for multiple time series
}
\examples{
data(voldata)
mosum_sub(voldata[,2:5], 1, 250)
}
