% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum_univariate.R
\name{mosum_univ}
\alias{mosum_univ}
\title{Detect change points in the autoregressive structure of multiple time series x with the MOSUM procedure, using AR(p) models for each channel.
Optionally, perform dimension reduction and use residuals from a global VAR model.}
\usage{
mosum_univ(
  x,
  p,
  G,
  method = c("Wald", "Score")[1],
  estim = c("DiagC", "DiagH")[1],
  varEstim = c("Local", "Global")[1],
  alpha = 0.05,
  criterion = c("eps", "eta")[1],
  nu = 0.25,
  rm_cross_terms = FALSE,
  do_bootstrap = FALSE,
  M = 1000,
  global_resids = FALSE
)
}
\arguments{
\item{x}{data matrix}

\item{p}{integer VAR model order}

\item{G}{integer MOSUM bandwidth}

\item{method}{string indicating which of `Wald` or `Score` to use}

\item{estim}{string estimation method, `DiagC` or `DiagH`}

\item{varEstim}{string variance estimation method, `Local` or `Global`}

\item{alpha}{Numeric significance level}

\item{criterion}{string location procedure, `eps` or `eta`}

\item{nu}{Numeric location procedure hyperparameter}

\item{rm_cross_terms}{Boolean, perform dimension reduction}

\item{do_bootstrap}{Boolean, determine threshold via multiplier bootstrap method}

\item{M}{integer number of simulations for bootstrap}

\item{global_resids}{Boolean, use residuals from full VAR model}
}
\value{
list containing Boolean test outcome `Reject`, Numeric rejection threshold `Threshold`, 
 Numeric vector of test statistic `mosum`, Integer vector of estimated change points `cps`, Plot `plot`, 
 String of input estimator `estim`
}
\description{
Detect change points in the autoregressive structure of multiple time series x with the MOSUM procedure, using AR(p) models for each channel.
Optionally, perform dimension reduction and use residuals from a global VAR model.
}
\examples{
data(voldata)
mosum_univ(voldata[,2:5], 1, 250)
}
