% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{MOSUMBS}
\alias{MOSUMBS}
\title{MOSUM Binary Segmentation procedure for multiple time series}
\usage{
MOSUMBS(
  x,
  p,
  G,
  estim = c("DiagC", "DiagH")[1],
  varEstim = c("Local", "Global")[1],
  alpha = 0.05,
  criterion = c("eps", "eta")[1],
  nu = 0.25
)
}
\arguments{
\item{x}{data matrix}

\item{p}{integer VAR model order}

\item{G}{integer MOSUM bandwidth}

\item{estim}{string estimation method}

\item{varEstim}{string variance estimation method}

\item{alpha}{Numeric significance level}

\item{criterion}{string location procedure}

\item{nu}{Numeric location procedure hyperparameter}
}
\value{
list containing Boolean test outcome `Reject`, Numeric rejection threshold `Threshold`, 
 Numeric vector of test statistic `mosum`, Integer vector of estimated change points `cps`, Plot `plot`, 
 String of input estimator `estim`
}
\description{
MOSUM Binary Segmentation procedure for multiple time series
}
\examples{
data(voldata)
mosum_sub(voldata[,2:5], 1, 250)
}
