% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum_lm.R
\name{mosumlm.bs}
\alias{mosumlm.bs}
\title{Segment data under a linear regression model using MOSUM-Binary Segmentation}
\usage{
mosumlm.bs(
  X,
  y,
  G = NULL,
  intercept = TRUE,
  max.iter = 10,
  method = c("Wald", "Score"),
  alpha = 0.05,
  criterion = c("eps", "eta"),
  nu = 0.25,
  thresh = NULL,
  do.plot = TRUE
)
}
\arguments{
\item{X}{matrix of covariates}

\item{y}{vector of responses}

\item{G}{integer MOSUM bandwidth}

\item{intercept}{include intercept in regression}

\item{max.iter}{maximum number of recursions in Binary Segmentation}

\item{method}{detector, one of \code{"Wald", "Score"}}

\item{alpha}{Numeric significance level}

\item{criterion}{location procedure, one of \code{"eps", "eta"}}

\item{nu}{Numeric location procedure hyperparameter}

\item{thresh}{rejection threshold; see reference for default}

\item{do.plot}{Boolean, return plot}
}
\value{
List containing 
\itemize{
  \item{\code{thresh}}{ input}
  \item{\code{mosum}}{ vector of mosum detector values}
  \item{\code{cps}}{ estimated change points}
  \item{\code{plot}}{ detector plot}
}
}
\description{
Segment data under a linear regression model using MOSUM-Binary Segmentation
}
\examples{
data(voldata)
mosumlm.bs(voldata[,3:5], voldata[,2])
}
