% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.R
\name{mosumvar.bs}
\alias{mosumvar.bs}
\title{Segment data under a VAR model with MOSUM-Binary Segmentation}
\usage{
mosumvar.bs(
  x,
  order,
  G = NULL,
  estim = c("C", "H"),
  var.estim = c("Local", "Global"),
  alpha = 0.05,
  criterion = c("eps", "eta"),
  nu = 0.25,
  max.iter = 3,
  do.plot = TRUE
)
}
\arguments{
\item{x}{data matrix}

\item{order}{integer VAR model order}

\item{G}{integer MOSUM bandwidth}

\item{estim}{string estimation method}

\item{var.estim}{string variance estimation method}

\item{alpha}{Numeric significance level}

\item{criterion}{string location procedure}

\item{nu}{Numeric location procedure hyperparameter}

\item{max.iter}{integer maximum number of splits}

\item{do.plot}{Boolean, return plot}
}
\value{
List containing
\itemize{ 
  \item{\code{mosum}}{ vector of mosum detector values}
  \item{\code{cps}}{ estimated change points}
  \item{\code{plot}}{ detector plot}
  \item{\code{estim}}{ input}
}
}
\description{
Segment data under a VAR model with MOSUM-Binary Segmentation
}
\examples{
data(voldata)
mosumvar.bs(voldata[,2:5], 1, 250)
}
