% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mosumvar.fit}
\alias{mosumvar.fit}
\title{Fit a piecewise VAR model to data}
\usage{
mosumvar.fit(x, cps, order = NULL, pen = log(nrow(x))^1.01)
}
\arguments{
\item{x}{data matrix}

\item{cps}{change point vector}

\item{order}{integer VAR model order (optional, uses AIC otherwise)}

\item{pen}{penalty scalar; defaults to sSIC with exponent `1.01`}
}
\value{
list of model list and cost
}
\description{
Fit a piecewise VAR model to data
}
\examples{
data(voldata)
run_mosum <- mosumvar(voldata[,2:5], 1, 250)
mosumvar.fit(voldata[,2:5],run_mosum$cps)
}
