% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiscale.R
\name{mosumvar.ms}
\alias{mosumvar.ms}
\title{Segment data under a VAR model with multiple bandwidths}
\usage{
mosumvar.ms(
  x,
  order,
  Gset = NULL,
  method = c("Wald", "Score", "BS"),
  estim = c("C", "H"),
  alpha = 0.05,
  criterion = c("eps", "eta"),
  nu = 0.25,
  max.iter = 3,
  do.plot = TRUE
)
}
\arguments{
\item{x}{data matrix}

\item{order}{integer VAR model order}

\item{Gset}{integer vector of MOSUM bandwidths; see reference for default}

\item{method}{detector, one of \code{"Wald", "Score","BS"}}

\item{estim}{estimator method, one of \code{"C", "H"}}

\item{alpha}{Numeric significance level}

\item{criterion}{string location procedure}

\item{nu}{Numeric location procedure hyperparameter}

\item{max.iter}{integer maximum number of splits (when \code{method = "BS"})}

\item{do.plot}{Boolean, return plot}
}
\value{
List containing
\itemize{  
  \item{\code{cps}}{ estimated change points}
  \item{\code{plot}}{ detector plot}
  \item{\code{estim}}{ input}
}
}
\description{
Segment data under a VAR model with multiple bandwidths
}
\examples{
data(voldata)
mosumvar.ms(voldata[1:3000,2:4], 1)
}
