% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mosumvar}
\alias{mosumvar}
\title{Detect change points in the autoregressive structure of multiple time series x with the MOSUM procedure}
\usage{
mosumvar(
  x,
  p,
  G,
  method = c("Wald", "Score")[1],
  estim = c("DiagC", "DiagH")[1],
  varEstim = c("Local", "Global")[1],
  alpha = 0.05,
  criterion = c("eps", "eta")[1],
  nu = 0.25,
  do_bootstrap = FALSE,
  M = 1000
)
}
\arguments{
\item{x}{data matrix}

\item{p}{integer VAR model order}

\item{G}{integer MOSUM bandwidth}

\item{method}{string indicating which of `Wald` or `Score` to use}

\item{estim}{string estimation method, `DiagC` or `DiagH`}

\item{varEstim}{string variance estimation method, `Local` or `Global`}

\item{alpha}{Numeric significance level}

\item{criterion}{string location procedure, `eps` or `eta`}

\item{nu}{Numeric location procedure hyperparameter}

\item{do_bootstrap}{Boolean, determine threshold via multiplier bootstrap method}

\item{M}{Integer; number of bootstrap replicates}
}
\value{
list containing Boolean test outcome `Reject`, Numeric rejection threshold `Threshold`, 
 Numeric vector of test statistic `mosum`, Integer vector of estimated changepoints `cps`, Plot `plot`, 
 String of input estimator `estim`
}
\description{
Detect change points in the autoregressive structure of multiple time series x with the MOSUM procedure
}
\examples{
data(voldata)
mosumvar(voldata[,2:5], 1, 250)
}
