% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{mosumvar.sub}
\alias{mosumvar.sub}
\title{Segment data under a VAR model on a coarse grid}
\usage{
mosumvar.sub(
  x,
  order,
  G = NULL,
  method = c("Wald", "Score"),
  estim = c("C", "H"),
  var.estim = c("Local", "Global"),
  kap = 0.1,
  alpha = 0.05,
  criterion = c("eps", "eta"),
  nu = 0.25
)
}
\arguments{
\item{x}{data matrix}

\item{order}{integer VAR model order}

\item{G}{integer MOSUM bandwidth; see reference for default}

\item{method}{detector, one of \code{"Wald", "Score"}}

\item{estim}{estimator method, one of \code{"C", "H"}}

\item{var.estim}{variance estimator method, one of \code{"Local", "Global"}}

\item{kap}{Numeric subsampling resolution constant}

\item{alpha}{Numeric significance level}

\item{criterion}{location procedure, one of \code{"eps", "eta"}}

\item{nu}{Numeric location procedure hyperparameter}
}
\value{
List containing
\itemize{
  \item{\code{thresh}}{ input}
  \item{\code{mosum}}{ vector of mosum detector values}
  \item{\code{cps}}{ estimated change points}
  \item{\code{plot}}{ detector plot}
  \item{\code{estim}}{ input}
}
}
\description{
Segment data under a VAR model on a coarse grid
}
\examples{
data(voldata)
mosumvar.sub(voldata[,2:5], 1, 250)
}
