% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum_univariate.R
\name{mosumvar.uni}
\alias{mosumvar.uni}
\title{Segment data under univariate AR models, optionally using dimension reduction}
\usage{
mosumvar.uni(
  x,
  order,
  G = NULL,
  method = c("Wald", "Score"),
  estim = c("C", "H"),
  var.estim = c("Local", "Global"),
  alpha = 0.05,
  criterion = c("eps", "eta"),
  nu = 0.25,
  rm.cross.terms = FALSE,
  do.bootstrap = FALSE,
  n.bootstrap = 1000,
  global.resids = FALSE,
  thresh = NULL,
  do.plot = TRUE
)
}
\arguments{
\item{x}{data matrix}

\item{order}{integer VAR model order}

\item{G}{integer MOSUM bandwidth; see reference for default}

\item{method}{detector, one of \code{"Wald", "Score"}}

\item{estim}{estimator method, one of \code{"C", "H"}}

\item{var.estim}{variance estimator method, one of \code{"Local", "Global"}}

\item{alpha}{Numeric significance level}

\item{criterion}{location procedure, one of \code{"eps", "eta"}}

\item{nu}{Numeric location procedure hyperparameter}

\item{rm.cross.terms}{Boolean, perform dimension reduction}

\item{do.bootstrap}{Boolean, determine threshold via bootstrap method}

\item{n.bootstrap}{integer number of simulations for bootstrap}

\item{global.resids}{Boolean, use residuals from full VAR model}

\item{thresh}{rejection threshold; see reference for default}

\item{do.plot}{Boolean, return plot}
}
\value{
List containing
\itemize{
  \item{\code{thresh}}{ input}
  \item{\code{mosum}}{ vector of mosum detector values}
  \item{\code{cps}}{ estimated change points}
  \item{\code{plot}}{ detector plot}
  \item{\code{estim}}{ input}
}
}
\description{
Segment data under univariate AR models, optionally using dimension reduction
}
\examples{
data(voldata)
mosumvar.uni(voldata[,2:5], 1, 250)
}
