% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreRcpp_Test.R
\name{score}
\alias{score}
\title{score test}
\usage{
score(
  x,
  order,
  G,
  Phi = NULL,
  eps = NULL,
  alpha = 0.05,
  estim = "C",
  var.estim = "Local",
  criterion = "eps",
  nu = 0.25,
  do.bootstrap = c(F, "multiplier", "regression")[1],
  n.bootstrap = 1000,
  thresh = NULL,
  do.plot = TRUE
)
}
\description{
score test
}
\keyword{internal}
