% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{VAR.sim}
\alias{VAR.sim}
\title{Simulate multiple time series from a VAR model}
\usage{
VAR.sim(
  n,
  mu = NULL,
  Sigma = NULL,
  coeffs,
  error_dist = c("normal", "t", "garch")[1],
  P1 = NULL,
  Q1 = NULL,
  df = 1
)
}
\arguments{
\item{n}{integer data length}

\item{mu}{Numeric vector of means, defaults to zero}

\item{Sigma}{error covariance matrix, defaults to identity}

\item{coeffs}{list or matrix of VAR coefficients}

\item{error_dist}{string of error distribution}

\item{P1}{Matrix for BEKK}

\item{Q1}{Matrix for BEKK}

\item{df}{Integer t-distribution degrees of freedom}
}
\value{
data frame of time series
}
\description{
Simulate multiple time series from a VAR model
}
\examples{
VAR.sim(1000)
}
