% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VAR_sim}
\alias{VAR_sim}
\title{Simulates VAR data
Internal function wrapped by ...}
\usage{
VAR_sim(n, mu, Sigma, coeffs, error_dist, P1, Q1, df = 1L)
}
\value{
Product of v1 and v2
}
\description{
Simulates VAR data
Internal function wrapped by ...
}
