% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{voldata}
\alias{voldata}
\title{Volatility data of five technology assets (IBM, AAPL, INTC, MSFT, ORCL), the S&P technology sector (XLK), and the S&P index (SP)}
\usage{
data(voldata)
}
\description{
Volatility data of five technology assets (IBM, AAPL, INTC, MSFT, ORCL), the S&P technology sector (XLK), and the S&P index (SP)
}
