% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald_RcppParallel.R
\name{wald}
\alias{wald}
\title{Wald test}
\usage{
wald(
  x,
  order,
  G,
  alpha = 0.05,
  estim = "DiagC",
  criterion = "eps",
  nu = 1/4,
  do.bootstrap = FALSE,
  n.bootstrap = 1000,
  thresh = NULL,
  do.plot = TRUE
)
}
\description{
Wald test
}
\keyword{internal}
