% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosum_univariate.R
\name{mosum_univ}
\alias{mosum_univ}
\title{MOSUM procedure using dimension reduction}
\usage{
mosum_univ(
  x,
  p,
  G,
  method = c("Wald", "Score")[1],
  estim = c("DiagC", "DiagH")[1],
  varEstim = c("Local", "Global")[1],
  alpha = 0.05,
  criterion = c("eps", "eta")[1],
  nu = 0.25,
  rm_cross_terms = F,
  do_bootstrap = c(F, "multiplier", "regression")[1],
  M = 1000,
  global_resids = F
)
}
\arguments{
\item{x}{data matrix}

\item{p}{integer VAR model order}

\item{G}{integer MOSUM bandwidth}

\item{method}{string indicating which of `Wald` or `Score` to use}

\item{estim}{string estimation method}

\item{varEstim}{string variance estimation method}

\item{alpha}{Numeric significance level}

\item{criterion}{string location procedure}

\item{nu}{Numeric location procedure hyperparameter}

\item{rm_cross_terms}{Boolean perform dimension reduction}

\item{do_bootstrap}{string threshold bootstrap method, `multiplier`}

\item{M}{integer number of simulations for `do_bootstrap`}

\item{global_resids}{Boolean use}
}
\value{
list containing Boolean test outcome `Reject`, Numeric rejection threshold `Threshold`, 
 Numeric vector of test statistic `mosum`, Integer vector of estimated change points `cps`, Plot `plot`, 
 String of input estimator `estim`
}
\description{
MOSUM procedure using dimension reduction
}
\examples{
data(voldata)
mosum_univ(voldata[,2:5], 1, 250)
}
