% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mosumvar}
\alias{mosumvar}
\title{MOSUM procedure for multiple time series}
\usage{
mosumvar(
  x,
  p,
  G,
  method = c("Wald", "Score")[1],
  estim = c("DiagC", "DiagH")[1],
  varEstim = c("Local", "Global")[1],
  alpha = 0.05,
  criterion = c("eps", "eta")[1],
  nu = 0.25,
  do_bootstrap = c(F, "multiplier", "regression")[1],
  M = 1000
)
}
\arguments{
\item{x}{data matrix}

\item{p}{integer VAR model order}

\item{G}{integer MOSUM bandwidth}

\item{method}{string indicating which of `Wald` or `Score` to use}

\item{estim}{string estimation method}

\item{varEstim}{string variance estimation method}

\item{alpha}{Numeric significance level}

\item{criterion}{string location procedure}

\item{nu}{Numeric location procedure hyperparameter}

\item{do_bootstrap}{Bootstrap procedure to use (For Score procedure only)}

\item{M}{Integer; number of bootstrap replicates}
}
\value{
list containing Boolean test outcome `Reject`, Numeric rejection threshold `Threshold`, 
 Numeric vector of test statistic `mosum`, Integer vector of estimated changepoints `cps`, Plot `plot`, 
 String of input estimator `estim`
}
\description{
MOSUM procedure for multiple time series
}
\examples{
data(voldata)
mosumvar(voldata[,2:5], 1, 250)
}
