% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{VAR.sim}
\alias{VAR.sim}
\title{Simulate multiple time series from a VAR model}
\usage{
VAR.sim(
  n,
  mu = NULL,
  Sigma = NULL,
  coeffs,
  error.dist = c("normal", "t", "garch"),
  P1 = NULL,
  Q1 = NULL,
  df = 3
)
}
\arguments{
\item{n}{integer data length}

\item{mu}{Numeric vector of means, defaults to zero}

\item{Sigma}{error covariance matrix, defaults to identity}

\item{coeffs}{list or matrix of VAR coefficients; model dimension and order are inferred from this}

\item{error.dist}{string for error distribution, one of \code{"normal", "t", "garch"}}

\item{P1}{Covariance matrix for BEKK garch(1)}

\item{Q1}{Autoregression matrix for BEKK garch(1)}

\item{df}{Integer degrees of freedom for t-distribution}
}
\value{
data frame of simulated time series
}
\description{
Simulate multiple time series from a VAR model
}
\examples{
A <- diag(0.7,4)
data <- VAR.sim(100, coeffs=A)
plot.ts(data)
}
